# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
...
info:
    description: '# Documentation for Product API'
    title: of Product API
    version: 1.0.0
paths:
...
    /products/{id}:
        delete:
            description: Deletes a product
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
produces:
    - application/json
...
schemes:
    - http
swagger: "2.0"